Change MarkDown formatting
The README file had some rendering issues with MarkDown, which made parts of the documentation difficult to read. Furthermore, some stylistic choices made the text difficult to read (e.g. copious use of the `---` operator, bold text instead of code text or not formatting code snippets as code).

I have attempted to make the text more easy on the eyes.